pyproject.toml
[project]
name = "asteroids"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pygame-ce>=2.5.5",
    "pygame-gui>=0.6.14",
    "pygame-pgu>=0.21",
]
